+++ packages/snap/src/handlers/clientRequest/validation.ts
@@ -221,6 +221,21 @@ export const OnConfirmUnstakeRequestStruct = object({
 });
 
 /**
+ * claimUnstakedTrx request validation.
+ */
+export const ClaimUnstakedTrxRequestParamsStruct = object({
+  fromAccountId: UuidStruct,
+  assetId: NativeCaipAssetTypeStruct,
+});
+
+export const ClaimUnstakedTrxRequestStruct = object({
+  jsonrpc: JsonRpcVersionStruct,
+  id: JsonRpcIdStruct,
+  method: literal(ClientRequestMethod.ClaimUnstakedTrx),
+  params: ClaimUnstakedTrxRequestParamsStruct,
+});
+
+/**
  * Parses a base64-encoded rewards message.
  * Expected format: 'rewards,{address},{timestamp}'
  *